Added jprof to tree.  Makefile won't work without a surrounding mozilla tree, but I included the output of make so you can see what needs to be done (shouldn't be that complicated to make a working makefile).
server.cc now has the jprof code it needs, but it's #ifdef JPROF'ed out.
To enable, compile jprof, then compile server.cc with -DJPROF and with
appropriate changes to -I and -L so that it can find jprof.h and libjprof.so
(which is must be linked against).  Then read the README.html to see how
to enable profiling, process the output, and read the reports.