Commented out _emanager->restore(...)
I don't understand why we restore the pixels 'behind' a pointer when the
next thing we do is to delete the whole area that pointer is in.

This greatly reduces the flickering of the mousecursor when moving a window.

I'm rather sure this does not introduce any new redraw bugs...